feat: Criando a service de empresas e corrigindo os testes

+++ src/main/java/br/com/lunix/model/entities/Jogo.java
@@ -51,7 +51,7 @@ public class Jogo {
 
     private double notaMedia = 0.0;
 
-    private double totalAvaliacoes = 0;
+    private int totalAvaliacoes = 0;
 
     @DBRef
     private Empresa empresa;
